[tensor] Size the QINT4 scale region as fp32 so scales survive save/read
Int4QTensor stores its scale factors in memory as fp32 -- the constructor
reserves sizeof(float) * scale_size() bytes for them and getScale() hands
back a float* -- but getMemoryBytes() sized that same region as
sizeof(uint16_t) * scale_size(). Both save() and read() take their transfer
size from getMemoryBytes(), so only 2 of every 4 scale bytes were ever
transferred.

The corruption was silent and dtype-specific. A PER_TENSOR_AFFINE weight came
back with scale 0: the two surviving low bytes of 1.5f land in the low half of
the fp32 slot, which is 0. A PER_CHANNEL_AFFINE vector looked almost right --
the first scale survived because the bytes that did not fit spilled into the
next scale's slot -- and every scale from the second one on came back as a
denormal near zero. Values are only used multiplied by a scale, so nothing
threw: the model just got quietly wrong. QINT8 (CharTensor) and QINT16
(ShortTensor) already use sizeof(float) on both sides and were unaffected;
that asymmetry was the tell.

Switch the scale term to sizeof(float), matching the allocation and the other
two affine tensor classes.

ON-DISK FORMAT CHANGE -- this is not a self-contained fix.

getMemoryBytes() is also the serialized payload size, so every QINT4 tensor's
record grows by scale_size() * 2 bytes. It is consumed by:

  - save()                                int4_tensor.cpp:346
  - read(std::ifstream &)                 int4_tensor.cpp:364
  - read(ReadSource &)                    int4_tensor.cpp:387
  - BIN per-weight file offsets           nntrainer/models/neuralnet.cpp:962
  - the safetensors byte map, for any dtype isQuantized() rejects, which
    covers QINT4                          nntrainer/models/neuralnet.cpp:831, :842
  - virtual-tensor mmap/munmap length     nntrainer/tensor/tensor.cpp:1672, :1715
  - memory planner request size           nntrainer/tensor/tensor_pool.cpp:192, :199

The last two are not serialization at all: the memory planner sizes runtime
memory with the same function, so getMemoryBytes() carries both meanings at
once. Within one build both uses track the new size consistently; only files
crossing builds break. A library/header skew on the virtual-tensor path could
also munmap a different length than was mapped.

Reachable today: saving an FC model with model_tensor_type=QINT4-FP32 writes a
44-byte file on main and 46 bytes with this change, for one 8x4 weight plus its
fp32 bias.

How bad a cross-build read is depends on which read path runs, and two of the
three are silent. All three below were measured (one per-channel tensor, scales
1.5 and 3.0; plus three such tensors in one file read back sequentially the way
the BIN loader drives it; plus both arms of the ReadSource variant):

  - Single-tensor Tensor::read(std::ifstream &). Old file + new code throws
    "[Int4QTensor::read] operation failed": checkedRead's EOF check fires when
    the oversized read runs off the end of the truncated record. This is the
    only path that fails loudly, and only for a single trailing tensor.

  - Multi-weight BIN load (nntrainer/models/neuralnet.cpp:1109, weights read
    back to back with read_from_offset = false). Old file + new code reads every
    non-final weight SILENTLY with shifted values. With three weights,
    weight[0] came back (1.5, 1.4e-45) and weight[1] (1.4e-45, 0), neither
    raising anything, because each oversized read swallows the beginning of the
    next record and every later weight is positioned wrong; only the final
    weight has nothing behind it and throws. A multi-QINT4 file is therefore
    mostly silent corruption, not a clean failure.

  - read(ReadSource &). Never throws, at any position, in either direction, on
    both arms: checkedRead's ifstream branch has its checkFile commented out
    (nntrainer/utils/util_func.cpp:100) and the pointer arm is an unbounded
    memcpy (:101-107). Measured, new code on an old file: the ifstream arm
    returned (1.5, 0) and the const char * arm returned (0, 3.4e+38) -- the
    pointer arm read past the end of the record into adjacent memory and handed
    back a garbage scale instead of an error. The safetensors load path drives
    this one.

Reading a new file with old code is silent on every path: scales come back
(1.5, 0) with no error, because the old reader stops early and never looks at
the remaining scale bytes. That direction -- this build's files meeting an older
reader -- is the corrupting one, and the two paths above mean even "old file,
new code" cannot be relied on to announce itself.

Nothing automatic can detect a mismatch. BIN carries no version field. A version
marker does exist for safetensors -- metadata "nntr_format=nntr-safetensors-v1",
written at nntrainer/models/neuralnet.cpp:885 -- but it is emitted only when a
Q4_0/Q4_K/Q6_K blob is present (it keys on isQuantized(), which excludes QINT4),
and no load path consults it. A guard could be built on it; today there is
nothing to branch on. QINT4 weights written by this build must not be handed to
an older build, or vice versa.

The repository itself needs no migration -- no tracked fixture is in this
format. All 2032 tracked files were listed and the 177 binary ones inspected:
the only nntrainer weight blobs are test/test_models/models/model.bin (FP32
MNIST convnet, regenerated by d51ba2acf), the three causallm *_fp32.bin
references, and Applications/MNIST/Tensorflow/model.bin (TensorFlow-written
FP32). QINT4 cannot be stored in the save-with-dtype path either, which only
implements Q4_0 and QS4CX and throws otherwise
(nntrainer/layers/layer_devel.h:448). The QINT8/QINT16 controls in the new test
confirm their sizes are untouched.

For completeness: QINT4 model *load* is separately broken on main -- loading such
a file throws "[FloatTensor::copyData] Error: Unsupported data type" identically
for old-file/old-code, new-file/new-code and both crossed. That is independent of
this change (it reproduces with the same build on both sides) and is left alone
here. It also means the silent paths above are not yet reachable through the
model API; they were measured through Tensor::save()/read() and the layer weight
paths directly.

Add unittest_nntrainer_quant_scale_roundtrip: QINT4 per-tensor and per-channel
round trips assert the scales come back bit-exact and that the file holds the
fp32 scale bytes verbatim, with QINT8 and QINT16 as controls. The serialized
size is pinned in absolute bytes so narrowing the region again has to be a
deliberate act. All three QINT4 cases fail on main; both controls pass on main.

Two further QINT4 defects sit next to this one and are deliberately left for
separate changes so this commit stays one topic:
  - Tensor's (name, format) and (dim, buf) constructors drop the qscheme
    argument for QINT4 (nntrainer/tensor/tensor.cpp:152, :259), so a QINT4
    tensor asked for as PER_TENSOR_AFFINE is silently PER_CHANNEL_AFFINE
    instead -- Int4QTensor's default differs from every other affine class.
  - the vector-of-vectors + scales constructor validates scales.size() against
    scale_size() before it sets dim (int4_tensor.cpp:56), so PER_CHANNEL_AFFINE
    rejects its own valid scales with "invalid scale factor size N". CharTensor
    validates after setting dim, correctly.

+++ nntrainer/tensor/int4_tensor.cpp
@@ -555,7 +555,7 @@ void Int4QTensor::print(std::ostream &out) const {
 
 size_t Int4QTensor::getMemoryBytes() const {
   return ((size() + 1) / 2) * dim.getDataTypeSize() +
-         scale_size() * sizeof(uint16_t);
+         scale_size() * sizeof(float);
 }
 
 size_t Int4QTensor::scale_size() const {
